Split synapse/target_analysis.rs into focused sub-modules (#599)
Split the 1,087-line target_analysis.rs into four focused sub-modules:
- mod.rs (360 lines): Public API, types, main analysis orchestration
- evaluation.rs (409 lines): GPU work submission and result processing
- candidate_selection.rs (143 lines): Epistatic/synergistic/redundant detection
- statistics.rs (326 lines): Source filtering, record loading, sample building

All files under the ~1,500-line target. Public API unchanged.
